Fixed some major issues and you can now save and load linked arenas

+++ src/belven/arena/rewardclasses/ExperienceReward.java
@@ -2,9 +2,9 @@ package belven.arena.rewardclasses;
 
 public class ExperienceReward extends Reward
 {
-    public int experience = 0;
+    public double experience = 0;
 
-    public ExperienceReward(int exp)
+    public ExperienceReward(double exp)
     {
         rewardType = RewardType.Experience;
         experience = exp;
